+++ Wordle.Api/Wordle.Api/Services/PlayerService.cs
@@ -1,16 +1,21 @@
 using Microsoft.EntityFrameworkCore;
 using Wordle.Api.Models;
+using Wordle.Api.Data;
+using Wordle.Api.DTOS;
 
 namespace Wordle.Api.Services;
 
 public class PlayerService
 {
-    public WordleDbContext Db { get; set; }
+    private readonly AppDbContext _context;
+
+    private static object updateLock = new();
+	private static object addLock = new();
     
-    public PlayerService(WordleDbContext db)
-    {
-        Db = db;
-    }
+    public PlayerService(AppDbContext context)
+	{
+		_context = context;
+	}
 
     public async Task<Player?> GetPlayer(string playerName)
     {
@@ -19,28 +24,40 @@ public class PlayerService
     
     public async Task<Player[]> GetTopPlayers(int numberOfPlayers)
     {
-        return await Db.Players.OrderByDescending(p => p.AverageAttempts).Take(numberOfPlayers).ToArrayAsync();
+        return await Db.Players.OrderBy(p => p.AverageAttempts).Take(numberOfPlayers).ToArrayAsync();
     }
     
-    public async Task AddPlayer(Player player)
+    public async Task AddPlayer(PlayerDTO player)
     {
-        
-        var existingPlayer = await Db.Players.FirstOrDefaultAsync(p => p.PlayerId == player.PlayerId);
-        if (existingPlayer != null)
-        {
-            Db.Entry(existingPlayer).CurrentValues.SetValues(player);
-        }
-        else
-        {
-            await Db.Players.AddAsync(player);
-        }
-        
-        await Db.SaveChangesAsync();
-        // if (!Db.Players.Contains(player))
-        // {
-        //     await Db.Players.AddAsync(player);
-        //     await Db.SaveChangesAsync();
-        // }
-        
+        Player? curPlayer = await _context.Players.FirstOrDefaultAsync(player => player.Name.Equals(request.Name));
+		if (curPlayer != null)
+		{
+			lock (updateLock)
+			{
+				double attempts = curPlayer.AverageAttempts * curPlayer.GameCount + request.AverageAttempts;
+				curPlayer.AverageAttempts = attempts / (curPlayer.GameCount + 1);
+                int seconds = curPlayer.AverageSecondsPerGame * curPlayer.GameCount + request.AverageSecondsPerGame;
+                curPlayer.AverageSecondsPerGame = seconds/ (curPlayer.GameCount + 1);
+                curPlayer.GameCount = curPlayer.GameCount + 1;
+				_context.SaveChanges();
+				return curPlayer;
+			}
+		}
+		else
+		{
+			lock (addLock)
+			{
+                Player newPlayer = new()
+                {
+                    AverageAttempts = request.AverageAttempts,
+                    GameCount = request.GameCount,
+                    Name = request.Name,
+                    AverageSecondsPerGame = request.AverageSecondsPerGame
+                };
+                _context.Players.Add(newPlayer);
+                _context.SaveChanges();
+                return newPlayer;
+			}
+		}
     }
 }
