+++ Wordle.Api/Wordle.Api/appsettings.json
@@ -7,7 +7,6 @@
   },
     "AllowedHosts": "*",
     "ConnectionStrings": {
-      //TODO
-        "DefaultConnection": "Data Source=(localdb)\\MSSQLLocalDB; Database=INSERT HERE; Trusted_Connection=True; Integrated Security=true; MultipleActiveResultSets=true"
+        "DefaultConnection": "Data Source=(localdb)\\MSSQLLocalDB; Database=wordledb; Trusted_Connection=True; Integrated Security=true; MultipleActiveResultSets=true"
     }
 }
